% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_colnames.R
\name{add_colnames}
\alias{add_colnames}
\title{Insert column names as new row}
\usage{
add_colnames(df, pos = 1, filter.columns = TRUE, wrap = FALSE)
}
\arguments{
\item{df}{data frame}

\item{pos}{position of new row}

\item{filter.columns}{do not insert column names starting with a dot}

\item{wrap}{if TRUE, column names are enclosed in braces}
}
\value{
data frame with character columns and one additional row
}
\description{
Insert column names as new row
}
