% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_coef.R
\name{format_coef}
\alias{format_coef}
\title{Format coefficients}
\usage{
format_coef(coef, stars, digits = 2, latex = FALSE)
}
\arguments{
\item{coef}{coefficient value}

\item{stars}{significance stars}

\item{digits}{number of decimal places (default=2)}

\item{latex}{format stars as math exponents? (default: FALSE)}
}
\value{
formatted coefficients as string
}
\description{
Format coefficients
}
