% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stars.R
\name{get_stars}
\alias{get_stars}
\title{Get significance stars from p values}
\usage{
get_stars(p, sig = c(0.1, 0.05, 0.01))
}
\arguments{
\item{p}{p values}

\item{sig}{cutoff values for significance stars}
}
\value{
stars
}
\description{
Get significance stars from p values
}
