% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltx_legend.R
\name{ltx_legend}
\alias{ltx_legend}
\title{Return legend for significance stars}
\usage{
ltx_legend(width, align = "r", size = "\\\\scriptsize",
  sig = c("0.01", "0.05", "0.10"))
}
\arguments{
\item{width}{number of columns to be spanned}

\item{align}{text alignment (l, c, r)}

\item{size}{text size (defaults to "\\scriptsize")}

\item{sig}{cutoff values for significance stars}
}
\value{
string
}
\description{
Return legend for significance stars
}
