% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltx_multi.R
\name{ltx_multi}
\alias{ltx_multi}
\title{Join table cells for LaTeX export}
\usage{
ltx_multi(df, start, end = NULL, align = "c")
}
\arguments{
\item{df}{data frame}

\item{start}{top-left coordinate (2-dim vector, or 4-dim vector)}

\item{end}{bottom-right coordinate (2-dim vector or NULL)}

\item{align}{horizontal alignment of multicol (defaults to "c")}
}
\value{
data frame
}
\description{
Join table cells for LaTeX export
}
\examples{
\dontrun{
ltx.multi(df, c(2,2), c(3,4))
}
}
