% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltx_sideways.R
\name{ltx_sideways}
\alias{ltx_sideways}
\title{Rotate LaTeX cell contents}
\usage{
ltx_sideways(df, row, col)
}
\arguments{
\item{df}{table data}

\item{row}{row number}

\item{col}{column number}
}
\value{
table data
}
\description{
Rotate LaTeX cell contents
}
\examples{
\dontrun{
ltx.sideways(df, "test", 1)
}
}
