% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_coef.R
\name{reg_coef}
\alias{reg_coef}
\title{Get regression coefficients and significance stars}
\usage{
reg_coef(reg, vcov = NULL, sig = c(0.1, 0.05, 0.01))
}
\arguments{
\item{reg}{regression object}

\item{vcov}{covariance matrix, defaults to OLS}

\item{sig}{cutoff values for significance stars}
}
\value{
data frame with coefficients and statistics
}
\description{
Get regression coefficients and significance stars
}
