% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_text.R
\name{write_text}
\alias{write_text}
\title{Append text to a file}
\usage{
write_text(file, ...)
}
\arguments{
\item{file}{text file (already existing)}

\item{...}{one or multiple strings (to be pasted and written)}
}
\description{
Append text to a file
}
\examples{
write_text("test.txt", "hello", "world", "\\n")
}
\keyword{internal}
